bench/requirements.txt
mysql-connector-python==8.4.0
matplotlib==3.7.5
pymongo==4.10.1
